debounce update
- moves debounce.js functionality into api.js
- updates files that used debounce.js to now use api.js for checking debounce (lmao just one)
- updated debounce to only be checked if it's a valid command in the first place
- discord.js updated to v14.15.3
