+++ package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@discordjs/opus": "^0.9.0",
     "@discordjs/voice": "^0.16.1",
-    "discord.js": "^14.14.1",
+    "discord.js": "^14.15.3",
     "dotenv": "^16.3.1",
     "ffmpeg-static": "^5.2.0",
     "sodium-native": "^4.0.4",
